[SERVER][REFACTOR]seperate modules
