+++ bchat-server/config/db.config.js
@@ -1,0 +1,13 @@
+require("dotenv").config();
+const mysql = require("mysql2/promise");
+
+const pool = mysql.createPool({
+  host: process.env.DB_HOST,
+  user: process.env.DB_USER,
+  password: process.env.DB_PASSWORD,
+  database: process.env.DB_NAME,
+  multipleStatements: true,
+  connectionLimit: 20,
+});
+
+module.exports = pool;
