+++ bchat-server/config/message-types.js
@@ -1,0 +1,11 @@
+const MESSAGE_TYPE = {
+  GET_CHANNELS: "get_channels",
+  CHANNEL_USERS: "channel_users",
+  JOIN_CHANNEL: "join_channel",
+  SEND_MESSAGE: "send_message",
+  RECEIVE_MESSAGE: "receive_message",
+  DELETE_MESSAGE: "delete_message",
+  LEAVE_CHANNEL: "leave_channel",
+};
+
+module.exports = MESSAGE_TYPE;
