--- bchat-server/db_connection.js
@@ -1,13 +1,0 @@
-const mysql = require("mysql");
-
-module.exports = (con) => {
-  con.connect(function (err) {
-    if (err) throw err;
-    console.log("Connected!");
-
-    con.query(sql, function (err, result) {
-      if (err) throw err;
-      console.log("Result: " + result);
-    });
-  });
-};
