--- bchat-server/index.html
@@ -1,92 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <style>
-      body {
-        margin: 0;
-        padding-bottom: 3rem;
-        font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-          Helvetica, Arial, sans-serif;
-      }
-
-      #msg-form {
-        background: rgba(0, 0, 0, 0.15);
-        padding: 0.25rem;
-        position: fixed;
-        bottom: 0;
-        left: 0;
-        right: 0;
-        display: flex;
-        height: 3rem;
-        box-sizing: border-box;
-        backdrop-filter: blur(10px);
-      }
-      #msg-input {
-        border: none;
-        padding: 0 1rem;
-        flex-grow: 1;
-        border-radius: 2rem;
-        margin: 0.25rem;
-      }
-      #msg-input:focus {
-        outline: none;
-      }
-      #msg-form > button {
-        background: #333;
-        border: none;
-        padding: 0 1rem;
-        margin: 0.25rem;
-        border-radius: 3px;
-        outline: none;
-        color: #fff;
-      }
-
-      #messages {
-        list-style-type: none;
-        margin: 0;
-        padding: 0;
-      }
-      #messages > li {
-        padding: 0.5rem 1rem;
-      }
-      #messages > li:nth-child(odd) {
-        background: #efefef;
-      }
-    </style>
-    <script src="/socket.io/socket.io.js"></script>
-
-    <title>Document</title>
-  </head>
-  <body>
-    <ul id="messages"></ul>
-    <form id="msg-form">
-      <input id="msg-input" autocomplete="off" />
-      <button type="submit">Send</button>
-    </form>
-  </body>
-  <script>
-    const socket = io();
-
-    const msgForm = document.getElementById("msg-form");
-    const msgInput = document.getElementById("msg-input");
-
-    msgForm.addEventListener("submit", (e) => {
-      e.preventDefault();
-      if (msgInput.value) {
-        console.log("msg", msgInput.value);
-        socket.emit("chat message", msgInput.value);
-        msgInput.value = "";
-      }
-    });
-
-    socket.on("chat message", (msg) => {
-      const item = document.createElement("li");
-      item.textContent = msg;
-      messages.appendChild(item);
-      window.scrollTo(0, document.body.scrollHeight);
-    });
-  </script>
-</html>
