+++ bchat-server/server.js
@@ -1,0 +1,7 @@
+require("dotenv").config();
+
+const server = require("./app");
+
+server.listen(process.env.PORT, () => {
+  console.log("server listening on port 4000");
+});
